sec-edgar-data/requirements.txt
psycopg2
requests
python-dotenv
